> moved !help and !ping to info category
